+++ demandSide/src/graphs/FirmsSalesProjection.java
@@ -1,6 +1,5 @@
 package graphs;
 
-import java.math.BigDecimal;
 
 import org.apache.commons.math3.util.FastMath;
 
@@ -40,8 +39,8 @@ public class FirmsSalesProjection {
 				qualityToCoord(firm.getQuality()));
 	}
 
-	private double priceToCoord(BigDecimal price) {
-		return (price.doubleValue() - Scale.getMinPrice())
+	private double priceToCoord(double price) {
+		return (price - Scale.getMinPrice())
 				/ (Scale.getMaxPrice() - Scale.getMinPrice()) * (MAX_X - MIN_X)
 				+ MIN_X;
 	}
@@ -53,9 +52,9 @@ public class FirmsSalesProjection {
 				 * (MAX_Y - MIN_Y) + MIN_Y;
 	}
 
-	private double qualityToCoord(BigDecimal quality) {
+	private double qualityToCoord(double quality) {
 		return MAX_Z
-				- ((quality.doubleValue() - Scale.getMinQuality())
+				- ((quality - Scale.getMinQuality())
 						/ (Scale.getMaxQuality() - Scale.getMinQuality())
 						* (MAX_Z - MIN_Z) + MIN_Z);
 	}
