+++ demandSide/src/optimalPrice/ToBeExpelled.java
@@ -1,6 +1,5 @@
 package optimalPrice;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 
 import firms.Firm;
@@ -8,9 +7,9 @@ import firms.Firm;
 public class ToBeExpelled {
 	Firm f;
 	// empty means any price expels the firm
-	Optional<BigDecimal> optPriceToBeExpelled;
+	Optional<Double> optPriceToBeExpelled;
 	
-	public BigDecimal getPriceToBeExpelled(){
+	public double getPriceToBeExpelled(){
 		assert optPriceToBeExpelled.isPresent();
 		return optPriceToBeExpelled.get();
 	}
